+++ src/onegov.file/onegov/file/tests/test_utils.py
@@ -1,7 +1,7 @@
 import onegov.file
 
 from onegov.core.utils import module_path
-from onegov.file.utils import content_type_from_fileobj
+from onegov.file.utils import content_type_from_fileobj, word_count
 
 
 def test_content_type_from_fileobj():
@@ -18,3 +18,9 @@ def test_content_type_from_fileobj():
         'application/vnd.openxmlformats-'
         'officedocument.wordprocessingml.document'
     )
+
+
+def test_word_count():
+    assert word_count("") == 0
+    assert word_count("foo bar") == 2
+    assert word_count("It's simple really") == 3